backend/api/services/firestore.py
from google.cloud import firestore
from api.config import settings
from typing import Dict, Any, Optional

# Firestore client initialization placeholder
def get_firestore_client() -> firestore.Client:
    """Get Firestore client"""
    return firestore.Client()

async def get_user_save_data(user_id: str) -> Optional[Dict[str, Any]]:
    """Get save data for a user"""
    db = get_firestore_client()
    doc_ref = db.collection("saves").document(user_id)
    doc = doc_ref.get()
    
    if doc.exists:
        return doc.to_dict()
    return None

async def save_user_data(user_id: str, save_data: Dict[str, Any]) -> bool:
    """Save user data to Firestore"""
    db = get_firestore_client()
    doc_ref = db.collection("saves").document(user_id)
    
    try:
        doc_ref.set(save_data)
        return True
    except Exception:
        return False

async def delete_user_save(user_id: str) -> bool:
    """Delete user save data"""
    db = get_firestore_client()
    doc_ref = db.collection("saves").document(user_id)
    
    try:
        doc_ref.delete()
        return True
    except Exception:
        return False
